pass Tone from a Sequencer to children WIP
